feat(charm): add juju secret suppport to share the protected nats connection url
This commit adds support for juju secrets to share the protcted nats url
containing the auth_token. The older way of sharing the url through
relation data still remains and is available to maintain backwards
compatibility. However we should depracate this when juju 3.0 becomes
the de-facto standard for all the charms.