feat(ci): add release jobs
This commit adds two release jobs for releasing the charms. One job is
trigger on a push to master which publishes the charm to latest/edge.
The other job works to promote this published charm revision to
latest/candidate or latest/stable. This job works with a manual trigger
and only gives two options to publish the charms in a sequenced manner.