fix(ci): move publishing from Github to Launchpad
This commit removes the publishing step from Github as we do not have
arm64 runners available on Github Actions. We need to build arm64
compatible charms which can be done on Launchpad and the publishing to
`latest/edge` can be done from there itself.
However, the promotion of the charm can still be done once uploaded to
charmhub. So we are keeping the promotion jobs as is.