fix(tests): Ignore a timing issue between `client` and `ca-client` relation
This commit fixes the tests for TLS where a timing issue can happen when
both the Client and the CA Client relation for the NATS charm are
scheduled. If the Client relation forms first a non-TLS url is shared
with the remote unit as NATS had been configured with it when it
started. But by the time the connection was made NATS was reconfigured
to use the TLS information.