feat(.gh): enable charm builds with charmcraftcache
This improves build efficiency when the GitHub cache expires after
7 days. Instead of pulling and rebuilding Python dependencies from
scratch leading to long build times—charm builds can now leverage the
daily updated pip wheel cache hosted in charmcraftcache-hub.

This is especially useful for charms that are not frequently updated
and rely on github action cache to speed up build.

NOTE: by default, it's not enabled until a PR to charmcraftcache-hub
to support build pip wheel for nats-operator and cache is available
from charmcraftcache-hub.