feat(relations): use dedicated secret label from the nats relation
Juju secrets are associated with the model rather than a specific unit.
This means that secrets management is handled at the model level.
When deploying the NATS in ha, to prevent any nats client from reading
the NATs url by using the same secret label and caused the URL to
be overridden in the end, use dedicated secret label instead.