fix(relations): correctly handle secret updates and emit events
The secret.set_content() does not actually refresh or reload the
secret content but create a new revision, which can result in reading
outdated data when the secret is newly created or updated. To address
this, we now listen to the secret_changed event to ensure the private
key is available in the secret_changed callback. And only trigger
the tls_config_ready event if the secret was changed was successful,
avoiding race conditions and state inconsistencies.