fix(deps): pin rpds-py to 0.24.0
jsonschema v4.24.0 pulls in rpds-py>=0.7.1, but the latest rpds-py
(0.25.1), which was just released on may 21, 2025 requires feature
`edition2024`, which isn't supported by the cargo 1.78.0 or 1.18.0.
Until now, it's unstable and only available in the rust nighlty build.

Hence pinning rpds-py to 0.24.0 ensures compatibility and allows
the charm to build successfully.