fix(tests): use nats charm from the store in test_upgrade_switch
When running the `test_upgrade_switch` test case with `--charm-channel`
parameter, we should install the charm from the store instead
of using a locally built charm. This ensures the publishing workflow
is validated end-to-end against the same artifact that is going to
be released.

The API call

    ops_test.model.applications[charm_name].refresh(
        channel=charm_channel, switch=...)

is buggy when switching to a different charm. In our case, refreshing
from the legacy `nats-charmers-nats` to `nats` leads to issues:

- On Juju 3.x, it upgrades to the wrong revision:

      preparing operation "upgrade to ch:amd64/jammy/nats-11" for nats/0:
      failed to download charm "ch:amd64/jammy/nats-11" from API server:
      download request with archiveSha256 length 0 not valid

  where revision 11 corresponds to `arm64/jammy`, not the `2/edge`
  release we requested.

- On Juju 2.9, it fails with:

      juju.errors.JujuError: : selecting releases: charm or bundle not
      found for channel "2/edge", base "amd64/ubuntu/22.04"

As a workaround, invoke `juju refresh` directly instead of using the
Python API.